core/test/com/g5/services/TransactionServiceTest.java
package com.g5.services;

import com.g5.entities.PaymentEntity;
import com.g5.entities.TransactionEntity;
import com.g5.types.Account;
import com.g5.types.Payment;
import com.g5.types.Transaction;
import java.math.BigDecimal;
import java.util.List;
import javax.ejb.embeddable.EJBContainer;
import javax.naming.Context;
import javax.naming.NamingException;
import javax.persistence.EntityManager;
import javax.persistence.EntityManagerFactory;
import javax.persistence.Persistence;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

public class TransactionServiceTest {

    private final static String TRANSACTION_SERVICE_LOCAL = "java:global/classes/TransactionService!com.g5.services.TransactionServiceLocal";
    private final static String PERSISTENCE_UNIT = "g5-non-jta";

    private static EJBContainer container;
    private static Context context;
    private EntityManagerFactory entityManagerFactory;
    private EntityManager entityManager;
    private TestData testData;

    @BeforeClass
    public static void setUpClass() {
        container = EJBContainer.createEJBContainer();
        context = container.getContext();
    }

    @AfterClass
    public static void tearDownClass() {
        container.close();
    }

    @Before
    public void setUp() {
        entityManagerFactory = Persistence.createEntityManagerFactory(PERSISTENCE_UNIT);
        entityManager = entityManagerFactory.createEntityManager();
        testData = TestData.reset(entityManager);
    }

    @After
    public void tearDown() {
        entityManager.clear();
        entityManager.close();
        entityManagerFactory.close();
    }

    private TransactionServiceLocal getTransactionService() throws NamingException {
        return (TransactionServiceLocal) context.lookup(TRANSACTION_SERVICE_LOCAL);
    }

    @Test
    public void testRollbackTransaction() throws Exception {
        System.out.println("remove");

        Account account = testData.getAccounts().get(0);
        Transaction transaction = testData.getTransactions().get(0);

        BigDecimal balance = account.getBalance();
        BigDecimal transactionValue = transaction.getValue();

        TransactionServiceLocal service = getTransactionService();

        long minTime = System.currentTimeMillis() / 1000 * 1000;

        service.rollbackTransaction(transaction.getId());

        long maxTime = (System.currentTimeMillis() / 1000 + 1) * 1000;

        List<TransactionEntity> transactions = entityManager.createNamedQuery("Transaction.findAll", TransactionEntity.class).getResultList();

        assertEquals(testData.getTransactions().size() + 1, transactions.size());

        Transaction rollbackTransaction = transactions.get(transactions.size() - 1);

        assertEquals(transaction.getAccount().getId(), rollbackTransaction.getAccount().getId());
        assertTrue(rollbackTransaction.getDate().getTime() >= minTime);
        assertTrue(rollbackTransaction.getDate().getTime() <= maxTime);
        assertEquals("Rollback - " + transaction.getDescription(), rollbackTransaction.getDescription());
        assertEquals(transaction.getValue().negate(), rollbackTransaction.getValue());

        entityManager.refresh(account);

        assertEquals(0, balance.subtract(transactionValue).compareTo(account.getBalance()));
    }

    @Test
    public void testDeposit() throws Exception {
        System.out.println("deposit");

        Account account = testData.getAccounts().get(0);

        TransactionServiceLocal service = getTransactionService();

        BigDecimal balance = account.getBalance();
        BigDecimal deposit = new BigDecimal(10);

        long minTime = System.currentTimeMillis() / 1000 * 1000;

        service.deposit(account.getId(), deposit);

        long maxTime = (System.currentTimeMillis() / 1000 + 1) * 1000;

        List<TransactionEntity> transactions = entityManager.createNamedQuery("Transaction.findAll", TransactionEntity.class).getResultList();

        assertEquals(testData.getTransactions().size() + 1, transactions.size());

        Transaction transaction = transactions.get(transactions.size() - 1);

        assertTrue(transaction.getDate().getTime() >= minTime);
        assertTrue(transaction.getDate().getTime() <= maxTime);
        assertEquals("Deposit", transaction.getDescription());
        assertEquals(0, deposit.compareTo(transaction.getValue()));

        entityManager.refresh(account);

        assertEquals(0, balance.add(deposit).compareTo(account.getBalance()));
    }

    @Test
    public void testWithdraw() throws Exception {
        System.out.println("withdraw");

        Account account = testData.getAccounts().get(0);

        TransactionServiceLocal service = getTransactionService();

        BigDecimal balance = account.getBalance();
        BigDecimal withdraw = new BigDecimal(10);

        long minTime = System.currentTimeMillis() / 1000 * 1000;

        service.withdraw(account.getId(), withdraw);

        long maxTime = (System.currentTimeMillis() / 1000 + 1) * 1000;

        List<TransactionEntity> transactions = entityManager.createNamedQuery("Transaction.findAll", TransactionEntity.class).getResultList();

        assertEquals(testData.getTransactions().size() + 1, transactions.size());

        Transaction transaction = transactions.get(transactions.size() - 1);

        assertTrue(transaction.getDate().getTime() >= minTime);
        assertTrue(transaction.getDate().getTime() <= maxTime);
        assertEquals("Withdraw", transaction.getDescription());
        assertEquals(0, withdraw.negate().compareTo(transaction.getValue()));

        entityManager.refresh(account);

        assertEquals(0, balance.subtract(withdraw).compareTo(account.getBalance()));
    }

    @Test
    public void testFindById() throws Exception {
        System.out.println("findById");

        Transaction transaction = testData.getTransactions().get(0);

        TransactionServiceLocal service = getTransactionService();

        Transaction result = service.findById(transaction.getId());

        entityManager.refresh(transaction);

        assertEquals(transaction.getId(), result.getId());
        // TODO: This assertion fails. result date is slightly older
        //assertEquals(transaction.getDate(), result.getDate());
        assertEquals(transaction.getDescription(), result.getDescription());
        assertEquals(0, transaction.getValue().compareTo(result.getValue()));
    }

    @Test
    public void testFindByAccountId() throws Exception {
        System.out.println("findByAccountId");

        Account account = testData.getAccounts().get(0);

        TransactionServiceLocal service = getTransactionService();

        List<Transaction> result = service.findByAccountId(account.getId());

        List<TransactionEntity> transactions = entityManager.createNamedQuery("Transaction.findByAccount", TransactionEntity.class).setParameter("account", account).getResultList();

        assertEquals(transactions.size(), result.size());

        for (int i = 0; i < result.size(); i++) {
            Transaction t1 = transactions.get(i);
            Transaction t2 = result.get(i);

            entityManager.refresh(t1);

            assertEquals(t1.getId(), t2.getId());
            // TODO: This assertion fails. t2 date is slightly older
            //assertEquals(t1.getDate(), t2.getDate());
            assertEquals(t1.getDescription(), t2.getDescription());
            assertEquals(0, t1.getValue().compareTo(t2.getValue()));
        }
    }

}
